+++ app/modules/admin/Module.php
@@ -11,6 +11,7 @@ class Module extends \yii\base\Module
 	 */
 	public $layout = 'main';
 
+
 	/**
 	 * @inheritdoc
 	 */