+++ app/modules/admin/assets/AdminltePluginsAssetBundle.php
@@ -9,11 +9,11 @@ class AdminltePluginsAssetBundle extends AssetBundle
 	public $sourcePath = '@vendor/almasaeed2010/adminlte/plugins';
 
 	public $js = [
-		'datatables/dataTables.bootstrap.min.js',
+		'datatables/dataTables.bootstrap.min.js'
 		// more plugin Js here
 	];
 	public $css = [
-		'datatables/dataTables.bootstrap.css',
+		'datatables/dataTables.bootstrap.css'
 		// more plugin CSS here
 	];
 	public $depends = [