+++ app/modules/admin/config/main.php
@@ -1,4 +1,11 @@
 <?php
 
-return [
+$config = [
+	'modules' => [
+		'rbac' => [
+			'class' => 'justcoded\yii2\rbac\Module'
+		],
+	],
 ];
+
+return $config;