+++ config/app-web.php
@@ -32,6 +32,9 @@ $config = [
 			'loginUrl' => ['auth/login'],
 			'enableAutoLogin' => true,
 		],
+		'authManager' => [
+			'class' => 'justcoded\yii2\rbac\components\DbManager',
+		],
 		'urlManager' => [
 			'enablePrettyUrl' => true,
 			'showScriptName' => false,
@@ -94,14 +97,14 @@ if (YII_ENV_DEV) {
 	$config['modules']['debug'] = [
 		'class' => 'yii\debug\Module',
 		// uncomment the following to add your IP if you are not connecting from localhost.
-		//'allowedIPs' => ['127.0.0.1', '::1'],
+		'allowedIPs' => ['*'],
 	];
 
 	$config['bootstrap'][]    = 'gii';
 	$config['modules']['gii'] = [
 		'class' => 'yii\gii\Module',
 		// uncomment the following to add your IP if you are not connecting from localhost.
-		//'allowedIPs' => ['127.0.0.1', '::1'],
+		'allowedIPs' => ['*'],
 	];
 }